Fin fold method; Fin ex 29
Folding is a little complicated to understand due to anonymous functions. Note that intellij does not infer data types for anonymous functions properly. In case of mapFold function with the anonymous function (a: Tree[B], b: Tree[B]) => Branch(a, b), it is CRUCIAL to specify the types for a and b (Tree[B[) as intellij interprets it as Leaf[A] or Branch[A] etc. Always be verbose / explicit with the data types in anonymous functions.